requirements.txt
Flask
psycopg2
python-dotenv
pypika
argon2-cffi
pyjwt